Removed extra dependency

+++ src/main/java/Main.java
@@ -1,21 +1,5 @@
-import java.io.BufferedReader;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Stack;
-import java.util.TreeSet;
-import java.util.zip.GZIPInputStream;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import gnu.trove.map.hash.THashMap;
 
@@ -63,7 +47,7 @@ public class Main {
 				if (estimatesNormalized.contains(key)) {
 					truePositives++;
 					estimatedCount = estimatesNormalized.get(key);
-					relative_error += Math.abs(trueCount - estimatedCount) / (double) trueCount;
+					relative_error += Math.abs(trueCount - estimatedCount) / trueCount;
 				}
 
 			}
